back/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Akenaide", email = "kevin.mansang@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "strawberry-graphql[fastapi]>=0.227.2",
    "uvicorn[standard]>=0.24.0.post1",
    "pydantic-settings>=2.2.1",
    "elasticsearch>=8.13.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "ipython>=8.23.0",
]
